+++ src/micromotion/__init__.py
@@ -128,7 +128,7 @@ from .spectral import (
     spectral_peak,
 )
 
-__version__ = "1.12.1"
+__version__ = "1.12.2"
 
 __all__ = [